requirements.txt
django_phonenumber_field==4.0.0
Django==2.2.6
dj_database_url==0.5.0
djangorestframework==3.11.0
python-decouple==3.3
